Create taxon name component
